harden stog_ocaml toplevel error reporting
Probably due to a development-setup level kind of bug,
Errors.report_error fails on my machine when called from stog_ocaml
toplevel error handling rountine. It is quite disturbing to have the
error handling routine to fail itself due to an error, and makes
"make doc" fail with an obscure error message.

This patch implements a workaround to catch and report errors in this
Error.report_error function in a quite detailed way, including
exception backtrace, in the hope to help users determine whether the
issue is within stog directly or is due to their particular setup
regarding exported compiler libraries.

Thanks to Gabriel Scherer for this patch.